requirements.txt
gradio
numpy
google-generativeai
python-dotenv
Pillow